+++ backend/canisters/group/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Short circuit without calling `ic0.time()` if there have been no updates ([#3539](https://github.com/open-ic/open-chat/pull/3539))
+- Short circuit query calls prior to calling `ic0.time()` where possible ([#3542](https://github.com/open-ic/open-chat/pull/3542))
 
 ## [[2.0.675](https://github.com/open-ic/open-chat/releases/tag/v2.0.675-group)] - 2023-04-29